% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assocref.R
\docType{data}
\name{assocref}
\alias{assocref}
\title{Insect host-parasite interactions}
\format{
A data.frame with 11005 observations of 5 variables
}
\usage{
data(assocref)
}
\description{
Edgelist of known associations between insect host (`Host`) and pathogen (`Pathogen`), and
associated reference and indexing values. The variables are as follows:
}
\details{
\itemize{
  \item ERNnvp. EDWIP record number 
  \item refCode. Reference code
  \item Reference. Actual citation
  \item Host. Host species
  \item Pathogen. Pathogen species
 }
}
\keyword{datasets}
