% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosts.R
\docType{data}
\name{hosts}
\alias{hosts}
\title{Detailed information on insect host species}
\format{
A data.frame with 4392 observations of 18 variables
}
\usage{
data(hosts)
}
\description{
Data on host species taxonomy, habitat, diet preferences, and distribution (in Canada)
}
\details{
\itemize{
  \item RecordNo. Sequence from 1:nrow(hosts)
  \item DateEntered. Date of initial data entry
  \item Habitat. Habitat of host
  \item HostSpecies. Host species
  \item HostOrder. Host order
  \item HostFamily. Host family
  \item Synonyms. Other names for the host species
  \item Food. What does the host eat?
  \item genYr. Number of generations per year. Can be <1, =1, >1, or some combination of these.
  \item CommonName. Host common name
  \item ProvinceI. Canadian provinces where host has been found.
  \item InsectStatus. Is the insect a pest, beneficial, endangered, unknown? Factor variable with
    7 unique values
  \item ModificationDate. Modification date of entry
  \item InCanada. Citations for presence/absence of host in Canada. Numeric indices can be
    related to the citations in the `citations.rda` data file. `Y` and `N` 
    relate to presence and absence, respectively.
  \item ChangeSpeciesTo. 
  \item CommonNameOther. Other common names?
  \item Complete. Is this record complete?
  \item AdditionalReferences. Additional reference indices. 
  \item HostTaxID. Host NCBI ID number
  \item HostGenus. Host genus
  \item HostFamily. Host family
  \item HostOrder. Host order
  \item HostClass. Host class
 }
}
\keyword{datasets}
