% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hostTaxonomy.R
\docType{data}
\name{hostTaxonomy}
\alias{hostTaxonomy}
\title{Cached version of host taxonomy (see vignette for code to build from scratch)}
\format{
A data.frame
}
\usage{
data(hostTaxonomy)
}
\description{
Host taxonomic cached data from NCBI (see vignette for code to re-make)

\itemize{
  \item HostSpecies original host name from EDWIP
  \item HostTaxID NCBI taxonomic ID
  \item HostGenus host genus
  \item HostFamily host family
  \item HostOrder host order
  \item HostClass  host class
 }
}
\keyword{datasets}
