% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative.R
\docType{data}
\name{negative}
\alias{negative}
\title{Information on negative host-parasite interactions}
\format{
A data.frame with 529 observations of 12 variables
}
\usage{
data(negative)
}
\description{
Hosts were challenged with pathogen, and did not become infected. These are data on what
pathogens do not infect certain hosts.
}
\details{
\itemize{
  \item ERNntr. EDWIP record number
  \item PathogenSpecies. Pathogen species
  \item DateEntered. Date of initial data entry
  \item DateModified. Modification date of entry
  \item LogMaxDose. Dosage, in many different units
  \item HostStageTested. Host stage exposed to pathogen (e.g. Larvae, Nymph, Adult)
  \item HostSpecies. Host species examined
  \item Group. Pathogen group (e.g. viruses) 
  \item HighTaxon. General classification of pathogen (e.g. DNA virus)
  \item LowTaxon. More specific classification of pathogen (e.g. Baculoviridae)
  \item HostTaxID. Host NCBI ID number
  \item HostGenus. Host genus
  \item HostFamily. Host family
  \item HostOrder. Host order
  \item HostClass. Host class
  \item PathTaxID. Pathogen NCBI ID number
  \item PathGenus. Pathogen genus
  \item PathFamily. Pathogen family
  \item PathOrder. Pathogen order
  \item PathClass. Pathogen class
  \item PathKingdom. Pathogen kingdom
 }
}
\keyword{datasets}
