% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nematode.R
\docType{data}
\name{nematode}
\alias{nematode}
\title{Information on nematode parasite occurrences}
\format{
A data.frame with 234 observations of 15 variables
}
\usage{
data(nematode)
}
\description{
Host associations for nematode parasites
}
\details{
\itemize{
  \item ERNnem
  \item Host. Host species
  \item Nematode. Nematode parasite species
  \item NematodeOrder. Nematode parasite order
  \item NematodeFamily. Nematode parasite family
  \item NematodeStrain. Nematode parasite strain
  \item StageInfected. Host stage infected
  \item TissueInfected. Host tissue infected
  \item FieldOrLab. Was this a field or lab tested association?
  \item Country. What country did the interaction occur in?
  \item SoilType. Type of soil where interaction was observed
  \item AssociatedBacterium. Associated bacterium.
  \item IntermediateHost. Is there an intermediate host present?
  \item CreationDate. Date of initial data entry
  \item ModificationDate. Modification date of entry
 }
}
\keyword{datasets}
