% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{noassref}
\alias{noassref}
\title{noassref}
\format{
A data.frame
}
\usage{
data(noassref)
}
\description{
These data are included because they were originally part of the EDWIP data. 
We caution the user in using these data, as they do not have associated 
metadata that the other records have.
}
\details{
\itemize{
  \item RefCode. Index of reference 
  \item ERNntr. EDWIP record number
  \item HostSpecies. Host species
  \item PathogenSpecies. Pathogen species
  \item Citation. Citation for host-pathogen record
  \item HostTaxID. Host NCBI ID number 
  \item HostGenus. Host genus 
  \item HostFamily. Host family 
  \item HostOrder. Host order 
  \item HostClass. Host class 
  \item PathTaxID. Pathogen NCBI ID number 
  \item PathGenus. Pathogen genus 
  \item PathFamily. Pathogen family 
  \item PathOrder. Pathogen order 
  \item PathClass. Pathogen class 
  \item PathKingdom. Pathogen kingdom 
 }
}
\keyword{datasets}
