% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvpassoc.R
\docType{data}
\name{nvpassoc}
\alias{nvpassoc}
\title{Information on non-viral pathogens of insect hosts}
\format{
A data.frame with 7164 observations
}
\usage{
data(nvpassoc)
}
\description{
Non-viral parasite (protozoan, fungi, bacteria) occurrences on host species. The variables are 
as follows:
}
\details{
\itemize{
  \item ERNnvp. EDWIP record number
  \item PathogenSpecies. Pathogen species
  \item Group. Pathogen group (e.g. Protozoa)
  \item HostSpecies. Host species examined
  \item HostStageTested. Host stage exposed to pathogen (e.g. Larvae, Nymph, Adult)
  \item HostTissueInfected. Host tissue infected
  \item FieldOrLab. Was this a field or lab tested association?
  \item Country. What country did the interaction occur in?
  \item IntermediateHost. Is there an intermediate host present?
  \item DateEntered. Date of initial data entry
  \item DateModified. Modification date of entry
  \item BiogeographicRegion. Biogeographic region (or some combination thereof)
  \item HostTaxID. Host NCBI ID number
  \item HostGenus. Host genus
  \item HostFamily. Host family
  \item HostOrder. Host order
  \item HostClass. Host class
  \item PathTaxID. Pathogen NCBI ID number
  \item PathGenus. Pathogen genus
  \item PathFamily. Pathogen family
  \item PathOrder. Pathogen order
  \item PathClass. Pathogen class
  \item PathKingdom. Pathogen kingdom
 }
}
\keyword{datasets}
