% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen.R
\docType{data}
\name{pathogen}
\alias{pathogen}
\title{Information on pathogen species in the database}
\format{
A data.frame with 2041 observations of 5 variables
}
\usage{
data(pathogen)
}
\description{
Information on taxonomy of parasites present in this database
}
\details{
\itemize{
  \item PathogenSpecies. Pathogen species.
  \item Group. Pathogen group (e.g. Protozoa)
  \item HighTaxon. General classification of pathogen (e.g. Microspora)
  \item LowTaxon. More specific classification of pathogen (e.g. Microsporea)
  \item AdditionalNotes. Some additional notes.
 }
}
\keyword{datasets}
