% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathTaxonomy.R
\docType{data}
\name{pathTaxonomy}
\alias{pathTaxonomy}
\title{Cached version of pathogen taxonomy (see vignette for code to build from scratch)}
\format{
A data.frame
}
\usage{
data(pathTaxonomy)
}
\description{
\itemize{
  \item PathTaxID NCBI taxonomic ID
  \item PathNCBIResolved is the pathogen found in NCBI
  \item PathGenus pathogen genus
  \item PathFamily pathogen family
  \item PathOrder pathogen order
  \item PathClass  pathogen class
  \item PathKingdom  pathogen class
 }
}
\keyword{datasets}
