% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viraref.R
\docType{data}
\name{viraref}
\alias{viraref}
\title{Citation information for host-virus interactions}
\format{
A data.frame with 2124 observations of 5 variables
}
\usage{
data(viraref)
}
\description{
Edgelist of known associations between insect host (`Host`) and viral pathogen (`Virus`),
and associated references and indexing values. These data are smaller than `assocref`, and
`citation`. The variables are as follows:
}
\details{
\itemize{
  \item RefCode. Index of reference obtained from `nematode` data frame
  \item Citation. Reference.
  \item ERNv. EDWIP record number
  \item Host. Host species
  \item Virus. Virus name
 }
}
\keyword{datasets}
