% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assocref.R
\docType{data}
\name{assocref}
\alias{assocref}
\title{Insect host-parasite interactions}
\format{
A data.frame with 11005 observations of 5 variables
}
\usage{
data(assocref)
}
\description{
Edgelist of known associations between insect host (`Host`) and pathogen (`Pathogen`), and
associated reference and indexing values. The variables are as follows:
}
\details{
\itemize{
  \item ERNnvp. EDWIP record number 
  \item refCode. Reference code
  \item Reference. Actual citation
  \item HostSpecies. Host species
  \item PathogenSpecies. Pathogen species
  \item HostTaxID. Host NCBI ID number
  \item HostGenus. Host genus
  \item HostFamily. Host family
  \item HostOrder. Host order
  \item HostClass. Host class
  \item PathTaxID. Pathogen NCBI ID number
  \item PathGenus. Pathogen genus
  \item PathFamily. Pathogen family
  \item PathOrder. Pathogen order
  \item PathClass. Pathogen class
  \item PathKingdom. Pathogen kingdom
 }
}
\keyword{datasets}
