% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hostTaxonomy.R
\docType{data}
\name{hostTaxonomy}
\alias{hostTaxonomy}
\title{Cached version of host taxonomy (see vignette for code to build from scratch)}
\format{
A data.frame
}
\usage{
data(hostTaxonomy)
}
\description{
Cached version of host taxonomy (see vignette for code to build from scratch)
}
\details{
\itemize{
  \item HostOriginal original host name from EDWIP
  \item HostTaxID NCBI taxonomic ID
  \item HostNCBIResolved is the host found in NCBI
  \item HostGenus host genus
  \item HostFamily host family
  \item HostOrder host order
  \item HostClass  host class
 }
}
\keyword{datasets}
