% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negative.R
\docType{data}
\name{negative}
\alias{negative}
\title{Information on negative host-parasite interactions}
\format{
A data.frame with 529 observations of 12 variables
}
\usage{
data(negative)
}
\description{
Hosts were challenged with pathogen, and did not become infected. These are data on what
pathogens do not infect certain hosts.
}
\details{
\itemize{
  \item ERNntr. EDWIP record number
  \item Pathogen. Pathogen species
  \item DateEntered. Date of initial data entry
  \item DateModified. Modification date of entry
  \item LogMaxDose. Dosage, in many different units
  \item HostStageTested. Host stage exposed to pathogen (e.g. Larvae, Nymph, Adult, or 
     Egg\Larva\Pupa) 
  \item Host. Host species examined
  \item Group. Pathogen group (e.g. viruses) 
  \item HighTaxon. General classification of pathogen (e.g. DNA virus)
  \item LowTaxon. More specific classification of pathogen (e.g. Baculoviridae)
  \item Order. Host taxonomic order.
  \item Family. Host taxonomic family.
 }
}
\keyword{datasets}
