% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nematode.R
\docType{data}
\name{nematode}
\alias{nematode}
\title{Information on nematode parasite occurrences}
\format{
A data.frame with 234 observations
}
\usage{
data(nematode)
}
\description{
Host associations for nematode parasites
}
\details{
\itemize{
  \item ERNnem.  EDWIP record identifier
  \item HostSpecies. Host species
  \item PathogenSpecies. Nematode parasite species
  \item PathogenStrain. Nematode parasite strain
  \item StageInfected. Host stage infected
  \item TissueInfected. Host tissue infected
  \item FieldOrLab. Was this a field or lab tested association?
  \item Country. What country did the interaction occur in?
  \item SoilType. Type of soil where interaction was observed
  \item AssociatedBacterium. Associated bacterium.
  \item IntermediateHost. Is there an intermediate host present?
  \item CreationDate. Date of initial data entry
  \item ModificationDate. Modification date of entry
  \item HostTaxID. Host NCBI ID number
  \item HostGenus. Host genus
  \item HostFamily. Host family
  \item HostOrder. Host order
  \item HostClass. Host class
  \item PathTaxID. Pathogen NCBI ID number
  \item PathGenus. Pathogen genus
  \item PathFamily. Pathogen family
  \item PathOrder. Pathogen order
  \item PathClass. Pathogen class
  \item PathKingdom. Pathogen kingdom
 }
}
\keyword{datasets}
