% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{new_asso}
\alias{new_asso}
\alias{newnema}
\title{new_asso}
\format{
A data.frame of not good records

A data.frame
}
\usage{
data(new_asso)

data(newnema)
}
\description{
\itemize{
  \item ERNnew
  \item HostSpecies
  \item HostOrder
  \item HostFamily
  \item HostHabitat
  \item HostFood
  \item HostGenYr
  \item PathSpecies
  \item PathGroup
  \item PathHighTaxon
  \item PathLowTaxon
  \item StageInf
  \item TissueInfected
  \item Field
  \item Country
  \item IntermediateHost
  \item Citation
  \item MoreInfo
  \item Who
  \item CreationDate
  \item ModificationDate
  \item StainFCB
 }

\itemize{
  \item ERNnem. EDWIP record number
  \item Host
  \item Nematode
  \item NemaOrder
  \item NemaFamily
  \item NemaStrain
  \item StageInfected
  \item TissueInfected
  \item FieldOrLab
  \item Country
  \item SoilType
  \item AssociatedBacterium
  \item IntermediateHost
  \item CreationDate
  \item ModificationDate
 }
}
\keyword{datasets}
