% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{noassref}
\alias{noassref}
\title{noassref}
\format{
A data.frame
}
\usage{
data(noassref)
}
\description{
\itemize{
  \item RefCode. Index of reference 
  \item ERNntr. EDWIP record number
  \item Host. Host species
  \item Pathogen. Pathogen species
  \item Citation. Citation for host-nematode record
 }
}
\keyword{datasets}
