% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen.R
\docType{data}
\name{pathogen}
\alias{pathogen}
\title{Information on pathogen species in the database}
\format{
A data.frame with 2041 observations
}
\usage{
data(pathogen)
}
\description{
Information on taxonomy of parasites present in this database
}
\details{
\itemize{
  \item PathogenSpecies. Pathogen species.
  \item Group. Pathogen group (e.g. Protozoa)
  \item AdditionalNotes. Some additional notes.
  \item PathTaxID. Pathogen NCBI ID number
  \item PathGenus. Pathogen genus
  \item PathFamily. Pathogen family
  \item PathOrder. Pathogen order
  \item PathClass. Pathogen class
  \item PathKingdom. Pathogen kingdom
 }
}
\keyword{datasets}
