% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viraref.R
\docType{data}
\name{viraref}
\alias{viraref}
\title{Citation information for host-virus interactions}
\format{
A data.frame with 2124 observations
}
\usage{
data(viraref)
}
\description{
Edgelist of known associations between insect host (`Host`) and viral pathogen (`Virus`),
and associated references and indexing values. These data are smaller than `assocref`, and
`citation`. The variables are as follows:
}
\details{
\itemize{
  \item RefCode. Index of reference obtained from `nematode` data frame
  \item Citation. Reference.
  \item ERNv. EDWIP record number
  \item HostSpecies. Host species
  \item PathogenSpecies. Virus name
  \item HostTaxID. Host NCBI ID number
  \item HostGenus. Host genus
  \item HostFamily. Host family
  \item HostOrder. Host order
  \item HostClass. Host class
  \item PathTaxID. Pathogen NCBI ID number
  \item PathGenus. Pathogen genus
  \item PathFamily. Pathogen family
  \item PathOrder. Pathogen order
  \item PathClass. Pathogen class
  \item PathKingdom. Pathogen kingdom
 }
}
\keyword{datasets}
